Icons hotfixes

+++ repository/database/cooldown.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import Column, String, Integer
+from sqlalchemy import Column, String, BigInteger
 
 from repository.database import database
 
@@ -10,4 +10,4 @@ class Cooldown(database.base):
 
     command_name = Column(String, primary_key=True, nullable=False)
     user_id = Column(String, primary_key=True, nullable=False)
-    timestamp = Column(Integer, nullable=False)
+    timestamp = Column(BigInteger, nullable=False)
